More fine grained control over file overwriting
Builder stays in control which files can be overwritten (depending on the builder). Given the following scenario, that a generator contains two builders, one allows overwriting the other prohibits it. There is no chance of realizing this with the current apporach. With the 'guard' operator, the generator can only set everything to non-overwriting, while individual builders will deny this behavior change.